move types out of components
