+++ src/components/DreamsTagChart.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import dreams from '../data/data.json'
 import BarChart from './charts/BarChart.vue'
-import { TagCount } from './types/types'
+import { TagCount } from '../types/types'
 import { normalize } from '../data/utils'
 
 const tagCounts: { [tagName: string]: number } = {}
